requirements.txt
cffi==1.13.1
cryptography==2.8
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
